fixed bug with decreasing custom column value

+++ src/fullscreen_display.c
@@ -66,7 +66,7 @@ void fullscreen_display(struct search* search){
 							case SHORTCUT_TYPE_CUSTOM_COLUMN_DECREASE:{
 								if(tui_options.shortcuts[j].type==COLUMN_TYPE_TEXT) break;
 								get_file_columns(search->output_ids.data[i]);
-								float value = sqlite3_column_int(get_file_columns_statement, NON_CUSTOM_FILE_COLUMNS+tui_options.shortcuts[j].id);
+								int value = sqlite3_column_int(get_file_columns_statement, NON_CUSTOM_FILE_COLUMNS+tui_options.shortcuts[j].id);
 								if(value>custom_columns[tui_options.shortcuts[j].id].lower_limit) value--;
 								set_custom_column_value(search->output_ids.data[i], tui_options.shortcuts[j].id, &value);
 								break;}
